+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - When clicking on a heading in the ToC, the cursor is now set to that heading
   and the editor is focused immediately
 - Fix recent documents items not being clickable in Linux
+- Fix drag & drop of files onto the editor
 
 ## Under the Hood